magento-research/pwa-tests#MQE-1382-WaitForReactPageLoad
- Created a separate Class for the custom PWA actions.
- Removed the custom PWA actions from MagentoWebDriver.
