+++ src/Magento/FunctionalTestingFramework/Module/MagentoWebDriver.php
@@ -380,48 +380,6 @@ class MagentoWebDriver extends WebDriver
     }
 
     /**
-     * Wait for a PWA Element to NOT be visible using JavaScript.
-     *
-     * @param null $selector
-     * @param null $timeout
-     * @throws \Exception
-     * @return void
-     */
-    public function waitForPwaElementNotVisible($selector, $timeout = null)
-    {
-        $timeout = $timeout ?? $this->_getConfig()['pageload_timeout'];
-
-        // Determine what type of Selector is used.
-        // Then use the correct JavaScript to locate the Element.
-        if (\Codeception\Util\Locator::isXPath($selector)) {
-            $this->waitForJS("return !document.evaluate(`$selector`, document);", $timeout);
-        } else {
-            $this->waitForJS("return !document.querySelector(`$selector`);", $timeout);
-        }
-    }
-
-    /**
-     * Wait for a PWA Element to be visible using JavaScript.
-     *
-     * @param null $selector
-     * @param null $timeout
-     * @throws \Exception
-     * @return void
-     */
-    public function waitForPwaElementVisible($selector, $timeout = null)
-    {
-        $timeout = $timeout ?? $this->_getConfig()['pageload_timeout'];
-
-        // Determine what type of Selector is used.
-        // Then use the correct JavaScript to locate the Element.
-        if (\Codeception\Util\Locator::isXPath($selector)) {
-            $this->waitForJS("return !!document && !!document.evaluate(`$selector`, document);", $timeout);
-        } else {
-            $this->waitForJS("return !!document && !!document.querySelector(`$selector`);", $timeout);
-        }
-    }
-
-    /**
      * Wait for all visible loading masks to disappear. Gets all elements by mask selector, then loops over them.
      *
      * @param integer $timeout
